requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
argcomplete==3.1.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
boto==2.49.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.82
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.27.0
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.2
cycler==0.11.0
decorator==4.4.2
easydict==1.10
etils==1.4.0
executing==1.2.0
fasteners==0.18
filelock==3.12.2
flatbuffers==2.0.7
flax==0.5.1
fonttools==4.41.1
frozenlist==1.4.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gast==0.4.0
gcs-oauth2-boto-plugin==3.0
gdown==4.7.1
google-apitools==0.5.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.56.2
gsutil==5.25
h5py==3.9.0
httplib2==0.20.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.5
importlib-metadata==6.8.0
importlib-resources==6.0.0
iopath==0.1.10
ipython==8.14.0
jax==0.4.9
jaxlib==0.4.9+cuda11.cudnn86
jedi==0.19.0
Jinja2==3.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.6
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.8.0.74
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.3.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
portalocker==2.7.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pyu2f==0.1.5
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rich==11.1.0
rsa==4.7.2
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.32.0
tensorstore==0.1.40
termcolor==2.3.0
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
